Moved ShopifyCli::Helpers::AdminAPI to ShopifyCli::AdminAPI - changed namespace references in populate/resource.rb and   tasks/schema.rb - updated test cases under populate, admin_api and schema
