+++ lib/shopify-cli/helpers.rb
@@ -1,7 +1,6 @@
 module ShopifyCli
   module Helpers
     autoload :AccessToken, 'shopify-cli/helpers/access_token'
-    autoload :AdminAPI, 'shopify-cli/helpers/admin_api'
     autoload :Async, 'shopify-cli/helpers/async'
     autoload :EnvFile, 'shopify-cli/helpers/env_file'
     autoload :FS, 'shopify-cli/helpers/fs'
