+++ lib/shopify-cli/tasks/schema.rb
@@ -5,7 +5,7 @@ module ShopifyCli
     class Schema < ShopifyCli::Task
       def call(ctx)
         unless Helpers::Store.exists?(:shopify_admin_schema)
-          schema = Helpers::AdminAPI.query(ctx, 'admin_introspection')
+          schema = AdminAPI.query(ctx, 'admin_introspection')
           Helpers::Store.set(shopify_admin_schema: JSON.dump(schema))
         end
         schema = JSON.parse(Helpers::Store.get(:shopify_admin_schema))
